Add test case that runs the migration checker (verify it doesn't break)
