+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = test_project.settings
+django_find_project = true
+python_paths = tests/.temptest
