+++ tox.ini
@@ -12,6 +12,8 @@ deps =
     dj20: Django>=2.0,<2.1
     dj111: Django>=1.11,<1.12
     pytest
+    pytest-django
+    pytest-pythonpath
 
 setenv =
     PYTHONPATH = {toxinidir}:{env:PYTHONPATH:}
